Update for alloc page64

+++ include/common.h
@@ -104,7 +104,7 @@ struct pcb_t
 	char path[100];
 	struct code_seg_t *code; // Code segment
 	addr_t regs[10];	 // Registers, store address of allocated regions
-	uint32_t pc;		 // Program pointer, point to the next instruction
+	uint64_t pc;		 // Program pointer, point to the next instruction
 #ifdef MLQ_SCHED
 	// Priority on execution (if supported), on-fly aka. changeable
 	// and this vale overwrites the default priority when it existed
@@ -120,7 +120,7 @@ struct pcb_t
 
 	struct krnl_t *krnl;	
 	struct page_table_t *page_table; // Page table
-	uint32_t bp;			 // Break pointer
+	uint64_t bp;			 // Break pointer
 };
 
 /* Kernel structure */
